+++ packages/core/handle/handle.js
@@ -6,6 +6,7 @@
  * function in the chain. If it returns `false` (or any falsy value like `null` or `undefined`),
  * the event handling chain stops at that input function.
  *
+ * Example:
  * ```
  * import {forKey, forward, handle, preventDefault} from '@enact/core/handle';
  *
@@ -30,6 +31,7 @@
  * Handlers can either be bound directly using the native `bind()` method or using the `bindAs()`
  * utility method that is appended to the handler.
  *
+ * Example:
  * ```
  * import {forKey, forward, handle, preventDefault} from '@enact/core/handle';
  * import React from 'react';
@@ -57,6 +59,22 @@
  * ```
  *
  * @module core/handle
+ * @exports adaptEvent
+ * @exports call
+ * @exports callOnEvent
+ * @exports forward
+ * @exports forwardWithPrevent
+ * @exports forEventProp
+ * @exports forKey
+ * @exports forKeyCode
+ * @exports forProp
+ * @exports handle
+ * @exports log
+ * @exports oneOf
+ * @exports preventDefault
+ * @exports returnsTrue
+ * @exports stop
+ * @exports stopImmediate
  */
 
 import cond from 'ramda/src/cond';
@@ -131,16 +149,18 @@ const decorateHandleFunction = (fn) => {
  * of the returned value.
  *
  * @method   handle
- * @memberof core/handle
- * @param    {...Function}  handlers List of handlers to process the event
+ * @param    {...Function}  handlers List of handlers to process the event.
+ *
  * @returns  {Function}	    A function that accepts an event which is dispatched to each of the
  *                          provided handlers.
+ * @memberof core/handle
+ * @public
  */
 const handle = function (...handlers) {
 	const h = makeHandler(handlers);
 
 	// In order to support binding either handle (handle.bind(this)) or a handler
-	// (a = handle(), a.bind(this)), we cache `this` here and use is as the fallback for props and
+	// (a = handle(), a.bind(this)), we cache `this` here and use it as the fallback for props and
 	// context if fn() doesn't have its own `this`.
 	const _outer = this;
 
@@ -190,18 +210,22 @@ const handle = function (...handlers) {
  * arguments are passed to both the condition function and the handler function. The value returned
  * from the handler is returned.
  *
+ * Example:
  * ```
  * const handler = oneOf(
  * 	[forKey('enter'), handleEnter],
  * 	[forKey('left'), handleLeft],
  * 	[forKey('right'), handleRight]
  * );
+ * ```
  *
  * @method   oneOf
- * @memberof core/handle
- * @param    {...Function}  handlers List of handlers to process the event
+ * @param    {...Function[]}  handlers List of conditions and handlers to process the event
+ *
  * @returns  {Function}	    A function that accepts an event which is dispatched to each of the
  *                          conditions and, if it passes, onto the provided handler.
+ * @memberof core/handle
+ * @public
  */
 const oneOf = handle.oneOf = function (...handlers) {
 	return handle.call(this, cond(handlers));
@@ -211,6 +235,7 @@ const oneOf = handle.oneOf = function (...handlers) {
  * A function that always returns `true`. Optionally accepts a `handler` function which is called
  * before returning `true`.
  *
+ * Example:
  * ```
  * // Used to coerce an existing function into a handler change
  * const coercedHandler = handle(
@@ -226,9 +251,11 @@ const oneOf = handle.oneOf = function (...handlers) {
  * ```
  *
  * @method   returnsTrue
- * @memberof core/handle
- * @param    {[Function]}  handler  Handler function called before returning `true`
+ * @param    {Function}  [handler]  Handler function called before returning `true`.
+ *
  * @returns  {Function}	   A function that returns true
+ * @memberof core/handle
+ * @public
  */
 const returnsTrue = handle.returnsTrue = function (handler) {
 	if (handler) {
@@ -243,8 +270,9 @@ const returnsTrue = handle.returnsTrue = function (handler) {
 };
 
 /**
- * Calls a named function on the event and returns `true`
+ * Calls a named function on the event and returns `true`.
  *
+ * Example:
  * ```
  * import {callOnEvent, handle} from '@enact/core/handle';
  *
@@ -255,10 +283,11 @@ const returnsTrue = handle.returnsTrue = function (handler) {
  * ```
  *
  * @method   callOnEvent
- * @memberof core/handle
- * @param    {String}     methodName  Name of the method to call on the event.
- * @param    {Object}     ev          Event
+ * @param    {String}     methodName  Name of the method to call on the event
+ * @param    {Object}     ev          Event payload
+ *
  * @returns  {Boolean}                Always returns `true`
+ * @memberof core/handle
  * @private
  */
 const callOnEvent = handle.callOnEvent = curry((methodName, ev) => {
@@ -275,6 +304,7 @@ const callOnEvent = handle.callOnEvent = curry((methodName, ev) => {
 /**
  * Allows handling to continue if the value of `prop` on the event strictly equals `value`
  *
+ * Example:
  * ```
  * import {forEventProp, handle} from '@enact/core/handle';
  *
@@ -285,11 +315,13 @@ const callOnEvent = handle.callOnEvent = curry((methodName, ev) => {
  * ```
  *
  * @method   forEventProp
- * @memberof core/handle
  * @param    {String}	   prop   Name of property on event
  * @param    {*}           value  Value of property
- * @param    {Object}      ev     Event
+ * @param    {Object}      ev     Event payload
+ *
  * @returns  {Boolean}            Returns `true` if `prop` on `event` strictly equals `value`
+ * @memberof core/handle
+ * @public
  */
 const forEventProp = handle.forEventProp = curry((prop, value, ev) => {
 	return ev[prop] === value;
@@ -300,6 +332,7 @@ const forEventProp = handle.forEventProp = curry((prop, value, ev) => {
  * is not a function, it is ignored. The return value of the forwarded function is ignored and
  * `true` is always returned instead.
  *
+ * Example:
  * ```
  * import {forward, handle} from '@enact/core/handle';
  *
@@ -310,11 +343,13 @@ const forEventProp = handle.forEventProp = curry((prop, value, ev) => {
  * ```
  *
  * @method   forward
- * @memberof core/handle
  * @param    {String}    name   Name of method on the `props`
- * @param    {Object}    ev     Event
+ * @param    {Object}    ev     Event payload
  * @param    {Object}    props  Props object
+ *
  * @returns  {Boolean}          Always returns `true`
+ * @memberof core/handle
+ * @public
  */
 const forward = handle.forward = curry(named((name, ev, props) => {
 	const fn = props && props[name];
@@ -328,6 +363,7 @@ const forward = handle.forward = curry(named((name, ev, props) => {
 /**
  * Calls `event.preventDefault()` and returns `true`.
  *
+ * Example:
  * ```
  * import {handle, preventDefault} from '@enact/core/handle';
  *
@@ -338,9 +374,11 @@ const forward = handle.forward = curry(named((name, ev, props) => {
  * ```
  *
  * @method   preventDefault
- * @memberof core/handle
- * @param    {Object}        ev  Event
+ * @param    {Object}        ev  Event payload
+ *
  * @returns  {Boolean}           Always returns `true`
+ * @memberof core/handle
+ * @public
  */
 const preventDefault = handle.preventDefault = callOnEvent('preventDefault');
 
@@ -349,6 +387,7 @@ const preventDefault = handle.preventDefault = callOnEvent('preventDefault');
  * behavior. If the specified prop is `undefined` or is not a function, it is ignored. Returns
  * `false` when `event.preventDefault()` has been called in a handler.
  *
+ * Example:
  * ```
  * import {forwardWithPrevent, handle} from '@enact/core/handle';
  *
@@ -359,12 +398,13 @@ const preventDefault = handle.preventDefault = callOnEvent('preventDefault');
  * ```
  *
  * @method   forwardWithPrevent
- * @private
- * @memberof core/handle
  * @param    {String}    name   Name of method on the `props`
- * @param    {Object}    ev     Event
+ * @param    {Object}    ev     Event payload
  * @param    {Object}    props  Props object
+ *
  * @returns  {Boolean}          Returns `true` if default action is prevented
+ * @memberof core/handle
+ * @private
  */
 const forwardWithPrevent = handle.forwardWithPrevent = curry(named((name, ev, props) => {
 	let prevented = false;
@@ -382,6 +422,7 @@ const forwardWithPrevent = handle.forwardWithPrevent = curry(named((name, ev, pr
 /**
  * Calls `event.stopPropagation()` and returns `true`
  *
+ * Example:
  * ```
  * import {handle, stop} from '@enact/core/handle';
  *
@@ -392,15 +433,18 @@ const forwardWithPrevent = handle.forwardWithPrevent = curry(named((name, ev, pr
  * ```
  *
  * @method   stop
- * @memberof core/handle
- * @param    {Object}   ev  Event
+ * @param    {Object}   ev  Event payload
+ *
  * @returns  {Boolean}      Always returns `true`
+ * @memberof core/handle
+ * @public
  */
 const stop = handle.stop = named(callOnEvent('stopPropagation'), 'stop');
 
 /**
  * Calls `event.stopImmediatePropagation()` and returns `true`
  *
+ * Example:
  * ```
  * import {handle, stopImmediate} from '@enact/core/handle';
  *
@@ -411,15 +455,18 @@ const stop = handle.stop = named(callOnEvent('stopPropagation'), 'stop');
  * ```
  *
  * @method   stopImmediate
- * @memberof core/handle
- * @param    {Object}       ev  Event
+ * @param    {Object}       ev  Event payload
+ *
  * @returns  {Boolean}          Always returns `true`
+ * @memberof core/handle
+ * @public
  */
 const stopImmediate = handle.stopImmediate = callOnEvent('stopImmediatePropagation');
 
 /**
  * Allows event handling to continue if `event.keyCode === value`.
  *
+ * Example:
  * ```
  * import {forKeyCode, handle} from '@enact/core/handle';
  *
@@ -430,10 +477,12 @@ const stopImmediate = handle.stopImmediate = callOnEvent('stopImmediatePropagati
  * ```
  *
  * @method   forKeyCode
- * @memberof core/handle
  * @param    {Number}    value  `keyCode` to test
- * @param    {Object}    ev     Event
+ * @param    {Object}    ev     Event payload
+ *
  * @returns  {Boolean}          Returns `true` if `event.keyCode` strictly equals `value`
+ * @memberof core/handle
+ * @public
  */
 const forKeyCode = handle.forKeyCode = forEventProp('keyCode');
 
@@ -441,6 +490,7 @@ const forKeyCode = handle.forKeyCode = forEventProp('keyCode');
  * Allows handling to continue if the event's keyCode is mapped to `name` within
  * {@link core/keymap}.
  *
+ * Example:
  * ```
  * import {forKey, handle} from '@enact/core/handle';
  *
@@ -450,12 +500,14 @@ const forKeyCode = handle.forKeyCode = forEventProp('keyCode');
  * );
  * ```
  *
- * @method   forKey
- * @memberof core/handle
- * @param    {String}    name   Name from {@link core/keymap}
- * @param    {Object}    ev     Event
- * @returns  {Boolean}          Returns `true` if `event.keyCode` is mapped to `name`
  * @see      core/keymap
+ * @method   forKey
+ * @param    {String}    name   Name from {@link core/keymap}
+ * @param    {Object}    ev     Event payload
+ *
+ * @returns  {Boolean}          Returns `true` if `event.keyCode` is mapped to `name`
+ * @memberof core/handle
+ * @public
  */
 const forKey = handle.forKey = curry((name, ev) => {
 	return is(name, ev.keyCode);
@@ -464,6 +516,7 @@ const forKey = handle.forKey = curry((name, ev) => {
 /**
  * Allows handling to continue if the value of `prop` on the props strictly equals `value`.
  *
+ * Example:
  * ```
  * import {forProp, handle} from '@enact/core/handle';
  *
@@ -474,12 +527,14 @@ const forKey = handle.forKey = curry((name, ev) => {
  * ```
  *
  * @method   forProp
- * @memberof core/handle
  * @param    {String}    prop   Name of property on props object
  * @param    {*}         value  Value of property
- * @param    {Object}    ev     Event
+ * @param    {Object}    ev     Event payload
  * @param    {Object}    props  Props object
- * @returns  {Boolean}          Event handler
+ *
+ * @returns  {Boolean}          `true` if the value of `props[prop]` strictly equals `value`
+ * @memberof core/handle
+ * @public
  */
 const forProp = handle.forProp = curry((prop, value, ev, props) => {
 	return props[prop] === value;
@@ -489,6 +544,7 @@ const forProp = handle.forProp = curry((prop, value, ev, props) => {
  * Logs the event, props, and context optionally preceded by a custom message. Will only log in
  * development mode.
  *
+ * Example:
  * ```
  * import {forProp, handle, log} from '@enact/core/handle';
  *
@@ -499,11 +555,13 @@ const forProp = handle.forProp = curry((prop, value, ev, props) => {
  * ```
  *
  * @method   log
- * @memberof core/handle
  * @param    {String}     message  Custom message
- * @param    {Object}     ev       Event
+ * @param    {Object}     ev       Event payload
  * @param    {...*}       [args]   Any args passed are logged
+ *
  * @returns  {Boolean}             Always returns `true`
+ * @memberof core/handle
+ * @public
  */
 const log = handle.log = curry((message, ev, ...args) => {
 	if (__DEV__) {
@@ -522,6 +580,7 @@ const log = handle.log = curry((message, ev, ...args) => {
  *
  * If the method does not exist or handle isn't bound to an instance, it returns `false`.
  *
+ * Example:
  * ```
  * import {call, handle, forProp} from '@enact/core/handle';
  *
@@ -544,10 +603,12 @@ const log = handle.log = curry((message, ev, ...args) => {
  * ```
  *
  * @method   call
- * @memberof core/handle
  * @param    {String}     method  Name of method
+ *
  * @returns  {Boolean}            Returns the value returned by `method`, or `false` if the method
  *                                does not exist
+ * @memberof core/handle
+ * @public
  */
 const call = function (method) {
 	return named(function (...args) {
@@ -566,6 +627,7 @@ const call = function (method) {
  * `adapter` is passed as the first argument to `handler` with the remaining arguments kept the
  * same. This is often useful to generate a custom event payload before forwarding on to a callback.
  *
+ * Example:
  * ```
  * import {adaptEvent, forward} from '@enact/core/handle';
  *
@@ -580,11 +642,13 @@ const call = function (method) {
  * ```
  *
  * @method   adaptEvent
- * @memberof core/handle
  * @param    {Function}  adapter  Function to adapt the event payload
  * @param    {Function}  handler  Handler to call with the new event payload
  * @param    {...*}      [args]   Additional args passed to both `adapter` and `handler`
+ *
  * @returns  {Object}             New event payload
+ * @memberof core/handle
+ * @public
  */
 const adaptEvent = handle.adaptEvent = curry(function (adapter, handler) {
 	return named(function (ev, ...args) {