+++ packages/ui/FloatingLayer/index.js
@@ -7,11 +7,10 @@
  */
 
 import {FloatingLayer, FloatingLayerBase} from './FloatingLayer';
-import {FloatingLayerDecorator, contextTypes} from './FloatingLayerDecorator';
+import {FloatingLayerDecorator} from './FloatingLayerDecorator';
 
 export default FloatingLayer;
 export {
-	contextTypes,
 	FloatingLayer,
 	FloatingLayerBase,
 	FloatingLayerDecorator