+++ packages/moonstone/CHANGELOG.md
@@ -7,7 +7,28 @@ The following is a curated list of changes in the Enact moonstone module, newest
 ### Removed
 
 - `moonstone/Button` built-in support for tooltips
-- `moonstone/TooltipDecorator` config property `tooltipDestinationProp` in order to prevent directing tooltip into a property
+
+### Changed
+
+- `moonstone/spinner` to blur Spotlight when the spinner is active
+### Fixed
+- `moonstone/Scroller.Scroller`, `moonstone/VirtualList.VirtualGridList`, and `moonstone/VirtualList.VirtualList` to handle direction, page up, and page down keys properly on page controls them when `focusableScrollbar` is false
+- `moonstone/Scroller.Scroller`, `moonstone/VirtualList.VirtualGridList`, and `moonstone/VirtualList.VirtualList` to handle a page up or down key in pointer mode
+- `moonstone/VideoPlayer.MediaControls` to correctly handle more button color when the prop is not specified
+
+## [2.0.0-beta.9] - 2018-07-02
+
+### Added
+
+- `moonstone/ContextualPopupDecorator` instance method `positionContextualPopup()`
+- `moonstone/MoonstoneDecorator` config property `disableFullscreen` to prevent the decorator from filling the entire screen
+- `moonstone/Scroller` prop `onUpdate`
+
+### Fixed
+
+- `moonstone/Scrollable` to update scroll properly on pointer click
+- `moonstone/TooltipDecorator` to prevent unnecessary re-renders when losing focus
+- `moonstone/TooltipDecorator` to not dismiss the tooltip on pointer click
 
 ## [2.0.0-beta.8] - 2018-06-25
 
@@ -95,6 +116,7 @@ The following is a curated list of changes in the Enact moonstone module, newest
 - `moonstone/Slider` to read when knob gains focus or for a change in value
 - `moonstone/Scroller` to not cut off Expandables when scrollbar appears
 - `moonstone/VideoPlayer` to correctly read out when play button is pressed
+- `moonstone/Divider` to always use a fixed height, regardless of locale
 
 ## [2.0.0-beta.4] - 2018-05-21